fix(core): read the three double player distances as f64, not u64
`goal.execute` failed at the executor's first contact with a live game:

    goal: game rejected the command: failed to parse players:
    [{"name":"client1",...,"resource_reach_distance":
      2.70000000000000017763568394002504646778106689453125,...}]

`FactorioPlayer` declared `resource_reach_distance: u64` and the game sends
2.7, so serde rejected it. Every real player failed to deserialize, which
means `RconActuator::new` could never be built and no plan could ever run.

Checked all six distances against `LuaControl` in Factorio 2.1.17's
`runtime-api.json` (`read_type`, since `type` is null), not just the one that
happened to show a fraction:

    build_distance          uint32  u32  ok
    reach_distance          uint32  u32  ok
    drop_item_distance      uint32  u32  ok
    item_pickup_distance    double  u64  -> f64
    loot_pickup_distance    double  u64  -> f64
    resource_reach_distance double  u64  -> f64

The other two doubles happened to arrive whole and parsed by luck. The docs
also warn `resource_reach_distance` is "max double when not a character or
player connected to a character", so an integer was wrong on range too.

`PlayerChangedDistanceEvent` carries the same six `LuaControl` distances and
gets the same widths; it loses `Hash`/`Eq`, which `f64` does not have and
which nothing used. The `as f64` casts in `rcon.rs` and `planner/src/state.rs`
are now redundant and are dropped -- no logic change.

This is a schema correction, not a regression: `FactorioPlayer` is the
response body of four `crates/server` routes, and the integer schema described
a value the game never sends, so those routes are equally broken today for any
connected player. The generated `app/src/models/types.ts` is unchanged --
TypeScript spells both `u64` and `f64` as `number`.

Covered by three tests built on the verbatim live payload, fractional digits
and all. A fixture with whole-number distances parses fine against the broken
integer types and so cannot detect this.

+++ crates/core/src/factorio/rcon.rs
@@ -595,7 +595,7 @@ impl FactorioRcon {
         let action_id: ActionId = *next_action_id;
         *next_action_id = (*next_action_id + 1) % 1000;
         drop(next_action_id);
-        let resource_reach_distance = player.resource_reach_distance as f64;
+        let resource_reach_distance = player.resource_reach_distance;
         let distance = calculate_distance(&player.position, position);
         drop(player); // wow, without this factorio (?) freezes (!)
         if distance > resource_reach_distance {
